flashrom-view: Make _fileSelector a property
Failing to do this causes the render() function in this view to see an
undefined _fileSelector property. This is because render() happens
first, then firstUpdated() runs and picks out the file selector from the
shadowDOM and then render() doesn't run again. Make the file selector a
property so that we render() at least once more and pick up the real
file selector element to pass to writeFlash().

Signed-off-by: Stephen Boyd <swboyd@chromium.org>